refactor: extract ChatCompanion component from react-surface for better code organization
