+++ packages/plugins/plugin-assistant/src/capabilities/react-surface.tsx
@@ -3,29 +3,27 @@
 //
 
 import { Effect } from 'effect';
-import React, { useCallback, useEffect, useRef } from 'react';
+import React, { useCallback } from 'react';
 
 import { Capabilities, contributes, createIntent, createSurface, useIntentDispatcher } from '@dxos/app-framework';
 import { Blueprint } from '@dxos/blueprints';
 import { getSpace } from '@dxos/client/echo';
 import { Sequence } from '@dxos/conductor';
 import { InvocationTraceContainer } from '@dxos/devtools';
-import { Filter, Obj, Query, Ref } from '@dxos/echo';
+import { Obj } from '@dxos/echo';
 import { SettingsStore } from '@dxos/local-storage';
 import { SpaceAction } from '@dxos/plugin-space/types';
-import { useQuery } from '@dxos/react-client/echo';
-import { useAsyncEffect } from '@dxos/react-ui';
 import { StackItem } from '@dxos/react-ui-stack';
 
 import {
   AssistantSettings,
   BlueprintContainer,
+  ChatCompanion,
   ChatContainer,
   ChatDialog,
   PromptSettings,
   SequenceContainer,
 } from '../components';
-import { type AiChatProcessor } from '../hooks';
 import { ASSISTANT_DIALOG, meta } from '../meta';
 import { Assistant, AssistantAction } from '../types';
 
@@ -45,7 +43,6 @@ export default () =>
         Obj.instanceOf(Assistant.Chat, data.subject) && data.variant !== 'assistant-chat',
       component: ({ data, role }) => <ChatContainer role={role} chat={data.subject} />,
     }),
-    // TODO(burdon): Factor out to component.
     createSurface({
       id: `${meta.id}/companion-chat`,
       role: 'article',
@@ -55,8 +52,6 @@ export default () =>
       component: ({ data, role }) => {
         const { dispatch } = useIntentDispatcher();
 
-        // TODO(burdon): How should we manage multiple companion chats?
-        // TODO(burdon): Garbage collection of queues?
         const handleChatCreate = useCallback(async () => {
           const space = getSpace(data.companionTo);
           if (!space) {
@@ -64,6 +59,8 @@ export default () =>
           }
 
           // NOTE: The plugin's graph builder is currently responsible for selecting the (last) companion chat.
+          // TODO(burdon): How should we manage multiple companion chats?
+          // TODO(burdon): Garbage collection of queues?
           await Effect.runPromise(
             Effect.gen(function* () {
               const { object } = yield* dispatch(createIntent(AssistantAction.CreateChat, { space }));
@@ -80,41 +77,7 @@ export default () =>
           );
         }, [dispatch, data]);
 
-        // Initialize companion chat if it doesn't exist.
-        // TODO(wittjosiah): Factor out to container.
-        const space = getSpace(data.companionTo);
-        const companions = useQuery(
-          space,
-          Query.select(Filter.ids(data.companionTo.id)).targetOf(Assistant.CompanionTo).source(),
-        );
-        useAsyncEffect(async () => {
-          if (companions.length > 0) {
-            return;
-          } else {
-            await handleChatCreate();
-          }
-        }, [companions, data.subject]);
-
-        const processorRef = useRef<AiChatProcessor>();
-        const processor = processorRef.current;
-        useEffect(() => {
-          if (!processor) {
-            return;
-          }
-
-          // TODO(burdon): Check if already bound.
-          if (Obj.instanceOf(Blueprint.Blueprint, data.companionTo)) {
-            void processor.context.bind({ blueprints: [Ref.make(data.companionTo)] });
-          } else {
-            void processor.context.bind({ objects: [Ref.make(data.companionTo)] });
-          }
-        }, [processor, data.companionTo]);
-
-        if (data.subject === 'assistant-chat') {
-          return null;
-        }
-
-        return <ChatContainer ref={processorRef} role={role} chat={data.subject} onChatCreate={handleChatCreate} />;
+        return <ChatCompanion role={role} data={data} onChatCreate={handleChatCreate} />;
       },
     }),
     createSurface({
