+++ packages/plugins/plugin-assistant/src/components/index.ts
@@ -11,6 +11,7 @@ export * from './TemplateEditor';
 export * from './Toolbox';
 
 export const BlueprintContainer = lazy(() => import('./BlueprintContainer'));
+export const ChatCompanion = lazy(() => import('./ChatCompanion'));
 export const ChatContainer = lazy(() => import('./ChatContainer'));
 export const ChatDialog = lazy(() => import('./ChatDialog'));
 export const SequenceContainer = lazy(() => import('./SequenceContainer'));
